EES-2369 Move `downloadFile` util into common utils
