--- src/explore-education-statistics-admin/src/services/utils/file/downloadFile.ts
@@ -1,11 +1,0 @@
-const downloadFile = (blob: Blob, fileName: string) => {
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement('a');
-  link.href = url;
-  link.setAttribute('download', fileName);
-  document.body.appendChild(link);
-  link.click();
-  link.remove();
-};
-
-export default downloadFile;
